restructured directory to adhere to good conventions, still need to employ css modules to encapsulate styles
