+++ frontend/src/components/Navbar.js
@@ -1,6 +1,6 @@
 import React from "react"
 import { Link, useNavigate } from 'react-router-dom'
-import { supabase } from "./Supabase"
+import { supabase } from "../services/Supabase"
 
 function Navbar() {
   const navigate = useNavigate()
