+++ frontend/src/pages/LoadingScreen.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
-import './LoadingScreen.css';
+import '../assets/styles/LoadingScreen.css';
 
 function LoadingScreen() {
   const text = "Journal more with us";
