+++ frontend/src/pages/UserPage.js
@@ -1,4 +1,4 @@
-import Navbar from './Navbar'
+import Navbar from '../components/Navbar'
 import { Outlet } from 'react-router-dom'
 export function UserPage() {
   return (
